combine display/unlock phrase view
