+++ extension/src/popup/metrics/views.ts
@@ -22,9 +22,6 @@ const routeToEventName = {
   [ROUTES.accountCreator]: METRIC_NAMES.viewAccountCreator,
   [ROUTES.recoverAccount]: METRIC_NAMES.viewRecoverAccount,
   [ROUTES.recoverAccountSuccess]: METRIC_NAMES.viewRecoverAccountSuccess,
-  [ROUTES.unlockBackupPhrase]: METRIC_NAMES.viewUnlockBackupPhrase,
-  [ROUTES.displayBackupPhrase]: METRIC_NAMES.viewDisplayBackupPhrase,
-  [ROUTES.unlockBackupPhrase]: METRIC_NAMES.viewUnlockBackupPhrase,
   [ROUTES.displayBackupPhrase]: METRIC_NAMES.viewDisplayBackupPhrase,
   [ROUTES.settings]: METRIC_NAMES.viewSettings,
   [ROUTES.preferences]: METRIC_NAMES.viewPreferences,
