+++ extension/src/popup/styles.scss
@@ -6,8 +6,4 @@
   --popup--height: 600px;
 
   --bottom-nav--height: 72px;
-
-  ::-webkit-scrollbar {
-    display: none;
-  }
 }
