+++ extension/src/popup/views/DisplayBackupPhrase/styles.scss
@@ -1,4 +1,8 @@
 .DisplayBackupPhrase {
+  &--input {
+    margin: 6.25rem 0 1rem 0;
+  }
+
   &--submit {
     margin-top: 2.375rem;
   }
