--- extension/src/popup/views/UnlockBackupPhrase/styles.scss
@@ -1,5 +1,0 @@
-.UnlockBackupPhrase {
-  &--input {
-    margin: 6.25rem 0 1rem 0;
-  }
-}
